refactor(occurrence): centralizando actions em um arquivo occurrences na pasta actions
